+++ RandomUtil.cpp
@@ -11,9 +11,9 @@ bool RandomUtil::generateEvent(double probability)
 	return randomDistribution(randomGenerator);
 }
 
-int RandomUtil::randomInRange(int min, int max)
+int RandomUtil::randomInRange(size_t min, size_t max)
 {
-	std::uniform_int_distribution<int> dist(min, max);	
+	std::uniform_int_distribution<size_t> dist(min, max);
 	return dist(randomGenerator);
 }
